zlibraryCrowler/requirements.txt
# Core web scraping and automation
selenium>=4.0.0
beautifulsoup4>=4.11.0
aiohttp>=3.8.0
webdriver-manager>=3.8.0

# Environment and data handling
python-dotenv>=0.19.0
requests>=2.28.0

# HTML parsing
lxml>=4.9.0

# Async support and file handling
asyncio
typing-extensions>=4.0.0
aiofiles>=23.1.0

# Optional: For better async performance
uvloop>=0.17.0; sys_platform != "win32"

# Development and testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
